Created MessageComposerContent to get rid of renderContent()
